+++ ll/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0] - 2021-05-08
 ### Added
 - Added `Display` for `SocketInterruptMask`.
 - Added `Display` for `SocketMode`.
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Changed the display formatting for `LinkStatus` and `SocketInterrupt`.
+
+### Fixed
+- Fixed `BufferSize::try_from` returning incorrect values for 4, 8, and 16 KB.
 
 ## [0.7.0] - 2021-02-23
 ### Changed
@@ -107,7 +110,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-01-02
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-ll-rs/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/newAM/w5500-ll-rs/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/newAM/w5500-ll-rs/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/newAM/w5500-ll-rs/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/newAM/w5500-ll-rs/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/newAM/w5500-ll-rs/compare/v0.6.0...v0.6.1